requirements.txt
argcomplete==1.10.0
asgiref==3.2.10
boto3==1.17.3
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
Django==3.1.1
django-cloudinary-storage==0.3.0
django-utils-six==2.0
djangorestframework==3.11.1
docx2txt==0.8
EasyProcess==0.3
EbookLib==0.17.1
entrypoint2==0.2.4
extract-msg==0.23.1
gunicorn==20.0.4
idna==2.10
IMAPClient==2.1.0
lxml==4.6.3
mss==6.1.0
olefile==0.46
pdfminer.six==20181108
Pillow==8.2.0
psycopg2==2.8.6
pycryptodome==3.10.1
python-dateutil==2.8.1
python-http-client==3.3.2
python-pptx==0.6.18
pytz==2020.1
requests==2.24.0
sendgrid==6.7.0
six
sortedcontainers==2.3.0
soupsieve==2.2.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
tzlocal==1.5.1
urllib3==1.25.10
whitenoise==5.2.0
drf-yasg==1.20.0